先提一版加了一些管理和处理以及修复漏洞还得让ai看看

+++ hall/src/main/java/hall/connect/room/AckRoomTableHandler.java
@@ -79,7 +79,7 @@ public class AckRoomTableHandler implements ConnectHandle {
 	private HallProto.AckLogin buildLoginResponse(User user, List<Long> tables, List<ModelProto.Room> rooms) {
 		HallProto.AckLogin.Builder builder = HallProto.AckLogin.newBuilder()
 				.setCert(ByteString.copyFromUtf8(user.getDeviceId()))
-				.setUserId(user.getUserId())
+				.setUserId((int) user.getUserId())
 				.setNickName(ByteString.copyFromUtf8(user.getNick()))
 				.addAllTables(tables)
 				.addAllRoomList(rooms);
